Go: fix const-correctness in char* input and freearg typemaps
The Go `in` typemap for `char*`, `char[ANY]`, `char[]` used the variable
name `swig_str` (reserved for globals) and wrote through `$1` directly
which fails when `$1` is `const char*` (e.g. via `typedef const char* T`).

Use a plain local `temp` variable for the allocation and assign to `$1`
after the copy and null-terminator write, so the typemap works for both
`char*` and `const char*`.

Also use `(void *)` cast instead of `(char *)` in the `freearg` typemap,
since `$1` may be `const char*` and the generated wrappers compile as
C++ (`.cxx`), which requires an explicit cast to `void*` for `free()`.

New test case `go_const_char_input` (renamed from `typedef_const_char`
to follow Go-only test naming) covers `typedef const char*`, bare
`const char*`, `const char[]`, and `const char[16]`.
