+++ Lib/go/go.swg
@@ -441,10 +441,10 @@ rvrdeleter.reset($1); %}
 	char *, char[ANY], char[]
 %{
   {
-    char *swig_str = (char *)malloc($input.n + 1);
-    memcpy(swig_str, $input.p, $input.n);
-    swig_str[$input.n] = '\0';
-    $1 = ($1_ltype)swig_str;
+    char *temp = (char *)malloc($input.n + 1);
+    memcpy(temp, $input.p, $input.n);
+    temp[$input.n] = '\0';
+    $1 = ($1_ltype)temp;
   }
 %}
 
@@ -458,7 +458,7 @@ rvrdeleter.reset($1); %}
 
 %typemap(freearg)
 	char *, char[ANY], char[]
-%{ free((char *)$1); %}
+%{ free((void *)$1); %}
 
 %typemap(freearg) char *&
 %{ free(temp$argnum); %}
